Fix linked badges
It turns out that the link set up on shields.io is embedded into the
badge image, hence as a user, one has to first click on the badge to go
to its image page and then on the image again to go to the linked page.
This is definitely note intuitive for a user and hence it's much better
to put the badge image into an `<img>` element which is then in the link
name component of a normal Markdown link and then one can link the badge
directly to the desired page via a normal Markdown link.